refactor(graph-embedding-util): extract fit/ submodules from fit/mod.rs
Pure move — fit/mod.rs trimmed, {config,projection,cell_cell,samplers}.rs split out; public API unchanged.
